Added mouseClick listener and functions to draw a new piece based upon the pointer's coordinates returned relative to the canvas
Used jQuery to get the canvas offset so that is can be used to translate
the returned coordinates into the grid system. Added checks to verify
that the coordinate is within the grid before attempting to addPiece to
the game board. Added extra elseif condition in drawPieces to clearRect
if the boardPosition == null.