Reorganized application structure and resolved a db error when deploying to a container
Additionally, added a temporary skip button to resolve a blocker where
the player cannot play. The resolution, per Othello's rules is to skip
and allow the other play to continue playing.
Created server, lib, and client folders to organize the files. Added
collections to lib\collections.js so that it will instantiate first on
the server.