Added basic player collection and stored the  turn state in MongoDB
Added playerTurn to MongoDB and rewrote the turn system on Client.js to
use this value to determine turn and player input lockout. Added a
Collection observer to control the animated scroll on Chat to hopefully
solve an issue where the chat window does not scroll to the true bottom.