Added first pass on basic sessions. Added an argument to drawPieces
pieceData is tied to chat through a unique identifier. The UID is stored
in sessionStorage and allows the user to refresh the page without
dropping the game. Currently, the second player must manually connect
using the correct UID.
The extra argument in drawPieces is to reduce the number of times
boardPosition is referenced.