Rewrote the construction of each match item in the lobby to fix a bug with the delegate method extending a click listener to undesired divs
The extra click listeners on the child divs would cause a new match to
be created when clicking on the padding of the .floatMiddle element.
The child divs were removed to prevent this. Also added extra
preventative logic to avoid exceptions being thrown when a client is
trying to read the entry from a match that does not exist.